add nonce grind prefix config option
