+++ crates/clementine-tx-sender/.env.txsender
@@ -51,6 +51,9 @@ TX_SENDER_POLL_DELAY_MS=30000
 # optional, default=derived from (finality_depth * 2 * 10min) / poll_delay_ms, min 1, this denotes maximum number of consecutive input-unspent check failures before timing out a tx, considering it impossible to be sent if its input is spent by someone other tx, its better for this to be unset unless there are issues
 #TX_SENDER_INPUT_UNSPENT_MAX_RETRIES=100
 
+# optional, default=[2,2]; JSON byte-array prefix for wtxid/nonce grind (examples: [2], [2,2], [0,1,2])
+#TX_SENDER_NONCE_GRIND_PREFIX=[2,2]
+
 # JSON-RPC server (only used when json-rpc feature is enabled)
 # optional, default=127.0.0.1 (must be 127.0.0.1 or 0.0.0.0)
 TX_SENDER_JSONRPC_BIND=0.0.0.0
